Refactor client logic into own type
