+++ transport/stream_conn.go
@@ -1,4 +1,4 @@
-package serial
+package transport
 
 import (
 	"bytes"
